refactor: api data class 수정
