+++ app/src/main/java/com/mashup/ipdam/data/Review.kt
@@ -3,7 +3,7 @@ package com.mashup.ipdam.data
 data class Review(
     val buildingName: String,
     val createdDate: String,
-    val id: String,
+    val id: Int,
     val images: List<String>,
     val bookmark: Boolean,
     val nickName: String,
