+++ app/src/main/java/com/mashup/ipdam/data/ReviewMarker.kt
@@ -1,8 +1,8 @@
 package com.mashup.ipdam.data
 
 data class ReviewMarker(
-    val count: String,
-    val id: String,
-    val latitude: String,
-    val longitude: String
+    val count: Int,
+    val id: Int,
+    val latitude: Double,
+    val longitude: Double
 )
