+++ app/src/main/java/com/mashup/ipdam/data/db/AppDatabase.kt
@@ -1,4 +1,4 @@
-package com.mashup.ipdam.data
+package com.mashup.ipdam.data.db
 
 import android.content.Context
 import androidx.room.Database
