+++ app/src/main/java/com/mashup/ipdam/di/DatabaseModule.kt
@@ -1,7 +1,7 @@
 package com.mashup.ipdam.di
 
 import android.content.Context
-import com.mashup.ipdam.data.AppDatabase
+import com.mashup.ipdam.data.db.AppDatabase
 import com.mashup.ipdam.ui.search.data.source.HistoryDao
 import dagger.Module
 import dagger.Provides
