+++ app/src/main/java/com/mashup/ipdam/ui/home/MockReview.kt
@@ -8,7 +8,7 @@ object MockReview {
             Review(
                 "서울빌딩",
                 "2020.01.20",
-                "dahyun",
+                1,
                 listOf(
                     "https://search.pstatic.net/common/?src=http%3A%2F%2Fblogfiles.naver.net%2FMjAyMTAyMTlfMTM4%2FMDAxNjEzNjYzMDgyMzUz.UmYCoopR8Q8rTZTztLy9dHtZ9ZPZCIhHc8F7aUeQn5og.KfjcQvyb-xKR1G_sb28Mv7E8qZ4yBDz-Rf65j55Qhw0g.JPEG.jskdms007%2FIMG_6114.jpg&type=a340",
                     "dd"
@@ -21,7 +21,7 @@ object MockReview {
             Review(
                 "서울빌딩",
                 "2020.01.20",
-                "dahyun",
+                2,
                 listOf(
                     "https://search.pstatic.net/common/?src=http%3A%2F%2Fblogfiles.naver.net%2FMjAyMTAxMDhfMTY2%2FMDAxNjEwMTE0ODU5OTA0.AQHbOYKQYzAqXjSEmCXjSlKSPe5D87aRv1fr_opYeOIg.AJbs2vOn5bhicqiHogYAF7MvY4rrwvK8FpOHEPJ7tgog.JPEG.psaa980%2F20210105%25A3%25DF114311.jpg&type=a340",
                     "dd"
@@ -34,7 +34,7 @@ object MockReview {
             Review(
                 "서울빌딩",
                 "2020.01.20",
-                "dahyun",
+                3,
                 listOf(
                     "https://search.pstatic.net/common/?src=http%3A%2F%2Fblogfiles.naver.net%2FMjAyMDEyMDFfNTYg%2FMDAxNjA2Nzk5MDE2MDUw.U3qreKXfTgkNnzaPzcCMGjJXv59kacodEEn374dX8sgg.6SCh_iWN6Xiq4trm4A2A0TZraZGZP9lzA3N-oOXigbgg.PNG.monk0531%2Fimage.png&type=a340",
                     "dd"
@@ -47,7 +47,7 @@ object MockReview {
             Review(
                 "보빌딩",
                 "2020.01.20",
-                "dahyun",
+                4,
                 listOf(
                     "https://search.pstatic.net/common/?src=http%3A%2F%2Fblogfiles.naver.net%2FMjAyMTAyMDRfMTYx%2FMDAxNjEyNDQzMzkxNzg4.U_ov3cutDLr5SOKQCuqvIWnjhtY1xnyHb6gn2-D6rgEg.hGHrcNgvx9hRTHGpNqLg2j2KBIwufGfv0biR4XCXmsMg.JPEG.vit4566%2FIMG_0387.jpg&type=a340",
                     "dd"
@@ -62,7 +62,7 @@ object MockReview {
         Review(
             "서울빌딩",
             "2020.01.20",
-            "dahyun",
+            5,
             listOf(
                 "https://search.pstatic.net/common/?src=http%3A%2F%2Fblogfiles.naver.net%2FMjAyMTAxMDhfMTY2%2FMDAxNjEwMTE0ODU5OTA0.AQHbOYKQYzAqXjSEmCXjSlKSPe5D87aRv1fr_opYeOIg.AJbs2vOn5bhicqiHogYAF7MvY4rrwvK8FpOHEPJ7tgog.JPEG.psaa980%2F20210105%25A3%25DF114311.jpg&type=a340",
                 "dd"
@@ -75,7 +75,7 @@ object MockReview {
         Review(
             "서울빌딩",
             "2020.01.20",
-            "dahyun",
+            6,
             listOf(
                 "https://search.pstatic.net/common/?src=http%3A%2F%2Fblogfiles.naver.net%2FMjAyMDEyMDFfNTYg%2FMDAxNjA2Nzk5MDE2MDUw.U3qreKXfTgkNnzaPzcCMGjJXv59kacodEEn374dX8sgg.6SCh_iWN6Xiq4trm4A2A0TZraZGZP9lzA3N-oOXigbgg.PNG.monk0531%2Fimage.png&type=a340",
                 "dd"
@@ -88,7 +88,7 @@ object MockReview {
         Review(
             "보빌딩",
             "2020.01.20",
-            "dahyun",
+            7,
             listOf(
                 "https://search.pstatic.net/common/?src=http%3A%2F%2Fblogfiles.naver.net%2FMjAyMTAyMDRfMTYx%2FMDAxNjEyNDQzMzkxNzg4.U_ov3cutDLr5SOKQCuqvIWnjhtY1xnyHb6gn2-D6rgEg.hGHrcNgvx9hRTHGpNqLg2j2KBIwufGfv0biR4XCXmsMg.JPEG.vit4566%2FIMG_0387.jpg&type=a340",
                 "dd"
@@ -101,7 +101,7 @@ object MockReview {
         Review(
             "서울빌딩",
             "2020.01.20",
-            "dahyun",
+            8,
             listOf(
                 "https://search.pstatic.net/common/?src=http%3A%2F%2Fblogfiles.naver.net%2FMjAyMTAyMTlfMTM4%2FMDAxNjEzNjYzMDgyMzUz.UmYCoopR8Q8rTZTztLy9dHtZ9ZPZCIhHc8F7aUeQn5og.KfjcQvyb-xKR1G_sb28Mv7E8qZ4yBDz-Rf65j55Qhw0g.JPEG.jskdms007%2FIMG_6114.jpg&type=a340",
                 "dd"
@@ -116,7 +116,7 @@ object MockReview {
         Review(
             "서울빌딩",
             "2020.01.20",
-            "dahyun",
+            9,
             listOf(
                 "https://search.pstatic.net/common/?src=http%3A%2F%2Fblogfiles.naver.net%2FMjAyMDEyMDFfNTYg%2FMDAxNjA2Nzk5MDE2MDUw.U3qreKXfTgkNnzaPzcCMGjJXv59kacodEEn374dX8sgg.6SCh_iWN6Xiq4trm4A2A0TZraZGZP9lzA3N-oOXigbgg.PNG.monk0531%2Fimage.png&type=a340",
                 "dd"
@@ -129,7 +129,7 @@ object MockReview {
         Review(
             "보빌딩",
             "2020.01.20",
-            "dahyun",
+            10,
             listOf(
                 "https://search.pstatic.net/common/?src=http%3A%2F%2Fblogfiles.naver.net%2FMjAyMTAyMDRfMTYx%2FMDAxNjEyNDQzMzkxNzg4.U_ov3cutDLr5SOKQCuqvIWnjhtY1xnyHb6gn2-D6rgEg.hGHrcNgvx9hRTHGpNqLg2j2KBIwufGfv0biR4XCXmsMg.JPEG.vit4566%2FIMG_0387.jpg&type=a340",
                 "dd"
@@ -142,7 +142,7 @@ object MockReview {
         Review(
             "서울빌딩",
             "2020.01.20",
-            "dahyun",
+            11,
             listOf(
                 "https://search.pstatic.net/common/?src=http%3A%2F%2Fblogfiles.naver.net%2FMjAyMTAyMTlfMTM4%2FMDAxNjEzNjYzMDgyMzUz.UmYCoopR8Q8rTZTztLy9dHtZ9ZPZCIhHc8F7aUeQn5og.KfjcQvyb-xKR1G_sb28Mv7E8qZ4yBDz-Rf65j55Qhw0g.JPEG.jskdms007%2FIMG_6114.jpg&type=a340",
                 "dd"
@@ -155,7 +155,7 @@ object MockReview {
         Review(
             "서울빌딩",
             "2020.01.20",
-            "dahyun",
+            12,
             listOf(
                 "https://search.pstatic.net/common/?src=http%3A%2F%2Fblogfiles.naver.net%2FMjAyMTAxMDhfMTY2%2FMDAxNjEwMTE0ODU5OTA0.AQHbOYKQYzAqXjSEmCXjSlKSPe5D87aRv1fr_opYeOIg.AJbs2vOn5bhicqiHogYAF7MvY4rrwvK8FpOHEPJ7tgog.JPEG.psaa980%2F20210105%25A3%25DF114311.jpg&type=a340",
                 "dd"
